多云平台总览财务明细数据更新

+++ manycloud-common/src/main/java/com/core/manycloudcommon/caller/Item/FinanceStatsItem.java
@@ -2,11 +2,17 @@ package com.core.manycloudcommon.caller.Item;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class FinanceStatsItem {
+    /** 时间 **/
     private String dateStr;
-    private Integer recharge;
-    private Integer consumption;
-    private Integer withdrawal;
+    /** 充值金额 */
+    private BigDecimal recharge;
+    /** 消费金额 */
+    private BigDecimal consumption;
+    /** 提现金额 */
+    private BigDecimal withdrawal;
 
 }
