+++ manycloud-common/src/main/java/com/core/manycloudcommon/caller/vo/FinanceStatsVO.java
@@ -2,15 +2,21 @@ package com.core.manycloudcommon.caller.vo;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.util.List;
 import java.util.Map;
 
 @Data
 public class FinanceStatsVO {
+    /** 时间 **/
     private List<String> dates;
-    private List<Integer> recharge;
-    private List<Integer> consumption;
-    private List<Integer> withdrawal;
+    /** 金额 **/
+    private List<BigDecimal> recharge;
+    /** 充值金额 **/
+    private List<BigDecimal> consumption;
+    /** 提现金额 **/
+    private List<BigDecimal> withdrawal;
+    /** 总金额 **/
     private Map<String, Integer> total;
 
 }
